requirements.txt
psutil
requests
yara-python
tabulate
colorama
tqdm